bot-starter/requirements.txt
aiohttp>=3.9
python-dotenv>=1.0
solders>=0.21
base58>=2.1
